signer API: refactor, split to three internal submodules
This could be split even further (separate the interfaces from the
provided implementations) but this is already useful as it shows
our abstractions are not leaky: there are no cyclical dependencies.
